+++ packages/terra-core-docs/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 * Added
   * Added visual label and associated with input field using 'for' and 'id'.
+  * Added 'ariaLabel' attribute for `terra-form-select` examples.
+
 ## 1.51.0 - (November 16, 2023)
 
 * Changed